+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Change Log
 Version 0.1.2 *(2017-x-y)*
 ----------------------------
 * Adds "shake" animations. ([#13](https://github.com/PSPDFKit-labs/VanGogh/issues/13))
+    * Adds `ShakeAnimations` class with `shake()`, `shakeQuickly()` and `shakeSlowly()` methods.
+* Adds "show as toast" animations. ([#15](https://github.com/PSPDFKit-labs/VanGogh/issues/15))
+    * Adds `ToastAnimations` class with `showAsToast()`, `showAsToastQuickly()` and `showAsToastSlowly()` methods.
 
 Version 0.1.1 *(2017-11-20)*
 ----------------------------